cleaaaan (2)/cleaaaan/backend/requirements.txt
flask
flask-cors
python-dotenv
requests
werkzeug
